refactor: refactoring image, ordertest
- 이미지 경로 수정
- 주문 테스트 수정
